Restructure copycat exercise into main
