+++ src/components/App/App.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import WebNavBar from "../Navbar/Navbar.jsx";
 import Profile from "../Profile/Profile.js";
+import CopyCatContainer from "../CopyCat/CopyCatContainer";
 import './App.css';
 
 export default function App() {
@@ -9,6 +10,7 @@ export default function App() {
       <WebNavBar />
       <h1>React Practice Projects</h1>
       <Profile />
+      <CopyCatContainer />
     </div>
   );
 }
