+++ src/components/Navbar/Navbar.css
@@ -4,7 +4,8 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: end;
+  justify-content: center;
   font-size: calc(10px + 2vmin);
   color: white;
+  
 }
